Closes #5 (as well as possible for now)
Unfortunately, due to the sandboxing provided by Chrome, the user will need to change the context of the dev console to that provided by the extension.  That's the best we can do for now, because Chrome does some very clever sandboxery that is usually completely invisible, but causes issues here.  See the documentation here: https://developer.chrome.com/extensions/content_scripts#execution-environment